% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manacher.R
\name{manacher}
\alias{manacher}
\title{Manacher's Algorithm}
\usage{
manacher(s)
}
\arguments{
\item{s}{A character string.}
}
\value{
A character string representing the longest palindromic substring.
}
\description{
Finds the longest palindromic substring in a given string using Manacher's algorithm.
}
