% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manacher_dna.R
\name{manacher_dna}
\alias{manacher_dna}
\title{Manacher's Algorithm for DNA}
\usage{
manacher_dna(s, dna = TRUE)
}
\arguments{
\item{s}{A character string representing a DNA sequence.}

\item{dna}{A logical value indicating whether to check against the reverse complement.}
}
\value{
A character string representing the longest palindromic substring.
}
\description{
Finds the longest palindromic substring in a given character sequence using Manacher's algorithm.
Optionally checks against the reverse complement (dna).
}
