% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reverse_complement_dna}
\alias{reverse_complement_dna}
\title{Reverse Complement of a DNA Sequence}
\usage{
reverse_complement_dna(dna)
}
\arguments{
\item{dna}{A character string representing a DNA sequence.}
}
\value{
A character string representing the reverse complement of the DNA sequence.
}
\description{
Computes the reverse complement of a given DNA sequence.
}
