% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manacher.R
\name{manacher_dna}
\alias{manacher_dna}
\title{Manacher's Algorithm for DNA}
\usage{
manacher_dna(s, reverse_complement = FALSE)
}
\arguments{
\item{s}{A character string representing a DNA sequence.}

\item{reverse_complement}{A logical value indicating whether to check against the reverse complement.}
}
\value{
A character string representing the longest palindromic substring.
}
\description{
Finds the longest palindromic substring in a given DNA sequence using Manacher's algorithm.
Optionally checks against the reverse complement.
}
